+++ src/deadlink/_cli/check.py
@@ -1,15 +1,20 @@
 import argparse
 import sys
 
-from ..__about__ import __version__
 from .._main import (
     categorize_urls,
-    filter_allow_ignore,
     find_files,
     find_urls,
+    is_allowed,
     print_to_screen,
     read_config,
 )
+
+try:
+    from importlib import metadata
+except ImportError:
+    # Python 3.7 and earlier
+    pass
 
 
 def check(argv=None):
@@ -20,31 +25,46 @@ def check(argv=None):
     # get non-hidden files in non-hidden directories
     files = find_files(args.paths)
 
+    d = read_config()
+
     # filter files by allow-ignore-lists
     allow_patterns = set() if args.allow_files is None else set(args.allow_files)
-    ignore_patterns = set() if args.ignore_files is None else set(args.ignore_files)
-
-    d = read_config()
     if "allow_files" in d:
         allow_patterns = allow_patterns.union(set(d["allow_files"]))
+    ignore_patterns = set() if args.ignore_files is None else set(args.ignore_files)
     if "ignore_files" in d:
         ignore_patterns = ignore_patterns.union(set(d["ignore_files"]))
 
-    files, ignored_files = filter_allow_ignore(files, allow_patterns, ignore_patterns)
+    num_files_before = len(files)
+    files = set(
+        filter(lambda item: is_allowed(item, allow_patterns, ignore_patterns), files)
+    )
+    num_ignored_files = num_files_before - len(files)
+
+    allow_patterns = set() if args.allow_urls is None else set(args.allow_urls)
+    if "allow_urls" in d:
+        allow_patterns = allow_patterns.union(set(d["allow_urls"]))
+    ignore_patterns = set() if args.ignore_urls is None else set(args.ignore_urls)
+    if "ignore_urls" in d:
+        ignore_patterns = ignore_patterns.union(set(d["ignore_urls"]))
 
     urls = find_urls(files)
-    urls, ignored_urls = filter_allow_ignore(
-        urls,
-        set() if args.allow_urls is None else set(args.allow_urls),
-        set() if args.ignore_urls is None else set(args.ignore_urls),
+
+    num_urls_before = len(urls)
+    urls = set(
+        filter(lambda item: is_allowed(item, allow_patterns, ignore_patterns), urls)
     )
+    num_ignored_urls = num_urls_before - len(urls)
 
     print(
         f"Found {len(urls)} unique URLs in {len(files)} files "
-        f"(ignored {len(ignored_files)} files, {len(ignored_urls)} URLs)"
+        f"(ignored {num_ignored_files} files, {num_ignored_urls} URLs)"
     )
     d = categorize_urls(
-        urls, args.timeout, args.max_connections, args.max_keepalive_connections
+        urls,
+        args.timeout,
+        args.max_connections,
+        args.max_keepalive_connections,
     )
 
     print_to_screen(d)
@@ -111,6 +131,11 @@ def _get_parser():
         nargs="+",
         help="ignore file names containing these strings (e.g., .svg)",
     )
+
+    try:
+        __version__ = metadata.version("deadlink")
+    except Exception:
+        __version__ = "unknown"
 
     __copyright__ = "Copyright (c) 2021 Nico Schlömer <nico.schloemer@gmail.com>"
     version_text = "\n".join(